renames package of block processor
