+++ engine/verification/assigner/blockconsumer/reader_test.go
@@ -1,4 +1,4 @@
-package processor
+package blockconsumer
 
 import (
 	"testing"
